==> 2_code_datenaufbereitung/2_Abfluss_preprocessing.ipynb <==
# Preprocessing der Abflussdaten

Die Abflussdaten kommen von Online-Datenportal des LUBW. <br>
https://udo.lubw.baden-wuerttemberg.de/public/processingChain?repositoryItemGlobalId=hydrologische_landespegel&conditionValuesSetHash=EAA0E0C&selector=hydrologische_landespegel&sourceOrderAsc=false&offset=0&limit=2147483647 <br>
In diesem Downloadformat enthalten sie viel mehr Information, als wir brauchen. <br>
Mit diesem Code wird der Dataframe in das benötigte Format gebracht.

# benötigte libraries laden
import pandas as pd
import os

# working directory setzen
# work_dir = r"C:/Users/henry/Documents/A git/projektstudie-murgtal/"  
# os.chdir(work_dir)

# Datensatz einlesen
df = pd.read_csv("https://raw.githubusercontent.com/emorl/projektstudie-murgtal/main/1_daten/discharge/Hydrologische-Landespegel-2025-12-22.csv",
                 skiprows=6, # ohne Beschreibung in den ersten 7 Zeilen
                 sep=",", # seperator
                 encoding="latin1" # erkennt Umlaute
                 )

# irrelevante Spalten entfernen
df = df[["Datum / Uhrzeit", "Wert"]]
# verbliebende Spalten umbenennen
df.rename(columns={"Datum / Uhrzeit": "date"}, inplace=True)
df.rename(columns={"Wert": "flow_m3_s"}, inplace=True)

# Datentypen anpassen

# Datum in datetime Format
df["date"] = pd.to_datetime(df["date"], format="%Y-%m-%d", errors="coerce")
# Werte als numerische Werte und Punkt statt Komma als Dezimaltrenner
df["flow_m3_s"] = pd.to_numeric(
    df["flow_m3_s"].str.replace(",", ".", regex=False),
    errors="coerce"
)

# check
df.head(), df.tail(), df.describe()

# lokal abspeichern
df.to_csv("C:/Users/henry/Documents/A git/projektstudie-murgtal/1_daten/discharge/discharge_clean.csv", index=False)
