==> 2_code_datenaufbereitung/5_ERA5_Abfluss_merge.ipynb <==
# Abfluss-, Temperatur- und Niederschlag zusammenführen


# benötigte Pakete laden
import pandas as pd
import os

# Daten einlesen
era5_tas = pd.read_csv("https://raw.githubusercontent.com/emorl/projektstudie-murgtal/main/1_daten/ERA5_temp/era5_tas_daily_spatial_mean_1981_2023.csv")
era5_prec = pd.read_csv("https://raw.githubusercontent.com/emorl/projektstudie-murgtal/main/1_daten/ERA5_precipitation/era5_prec_daily_spatial_mean_1981_2023.csv")
discharge = pd.read_csv("https://raw.githubusercontent.com/emorl/projektstudie-murgtal/main/1_daten/discharge/discharge_clean.csv")

# Sicherstellen, dass "date" als Datetime-Objekt vorliegt
for df in [era5_tas, era5_prec, discharge]:
    df["date"] = pd.to_datetime(df["date"])

# Zusammenführen über die gemeinsame "date"-Spalte
merged_df = era5_tas.merge(era5_prec, on="date") \
                      .merge(discharge, on="date") #\
                      #.merge(rcp26_pr, on="date") \
                      #.merge(rcp85_pr, on="date")

# lokal peichern als neue CSV
output_csv = "C:/Users/henry/Documents/A git/projektstudie-murgtal/1_daten/era5_tas_prec_discharge_merge.csv"
merged_df.to_csv(output_csv, index=False)
