==> 2_code_datenaufbereitung/9_RCP85_temp_preprocessing.ipynb <==
# Datenaufbereitung RCP8.5 Temperatur

###  Packete laden, Pfade definieren und Daten laden

# Benötigte Packete
import xarray as xr
import geopandas as gpd
import rioxarray
import numpy as np
import pandas as pd
from pyproj import Geod
from shapely.geometry import Point
import os
from glob import glob

# Dateipfade definieren
nc_dir = "https://raw.githubusercontent.com/emorl/projektstudie-murgtal/main/1_daten/rcp85temp/raw/"

# Shapefile Pfad
shp_path = shp_path= "https://raw.githubusercontent.com/emorl/projektstudie-murgtal/main/1_daten/Murgtal_boundingbox/Murgtal_boundingbox.shp"

#Hier muss der lokale Pfad definiert werden:
output_csv = "hier-lokalen-Pfad-definieren/1_daten/rcp85temp/murg_tas_rcp85_mean.csv"

# Laden der Shapefile für das Murgeinzugsgebiet, Shapefile in Koordinatensystem EPSG 4326 umwandeln
murg_shape = gpd.read_file(shp_path).to_crs("EPSG:4326")

#Liste erstellen für Tageswerte
all_data = []

# ALle Dateien sortiert nach Jahr
nc_files = sorted(glob(os.path.join(nc_dir, "*.nc")))


### Loop zur Datenaufbereitung aller .nc Daten
Eine Shapefile kann nicht direkt auf eine .nc Datei angewendet werden. Shapefile = Vektoren, NC-Datei = Gitterzellen --> Maske schaut welche Gitterpunkte betroffen sind, sie hat das gleiche Format wie die Klimadaten in der .nc Datei

for nc_path in nc_files:
    print("Verarbeite:", nc_path)
    
    # Datei öffnen
    tas_ds = xr.open_dataset(nc_path)
    tas = tas_ds["tas"]

    # 2D lat/lon extrahieren
    lats = tas_ds["lat"].values
    lons = tas_ds["lon"].values

    # Maske erstellen (Gitterzellen im Murggebiet)
    mask = np.zeros_like(lats, dtype=bool)
    for geom in murg_shape.geometry:
        mask |= np.vectorize(lambda x, y: geom.contains(Point(x, y)))(lons, lats)
    mask_da = xr.DataArray(mask, dims=("rlat", "rlon"))

    # Auf Murg beschränken
    tas_murg = tas.where(mask_da)

    # Mittelwert berechnen
    mean_temp_celsius = (tas_murg.mean(dim=["rlat", "rlon"]) - 273.15).rename("temp_day")

    # In DataFrame umwandeln
    df = mean_temp_celsius.to_dataframe().reset_index()
    df = df[["time", "temp_day"]].rename(columns={"time": "date"})

    all_data.append(df)

# Alle Daten zusammenführen
final_df = pd.concat(all_data, ignore_index=True)
final_df["temp_day"] = pd.to_numeric(final_df["temp_day"], errors="coerce")

# CSV speichern
final_df.to_csv(output_csv, index=False)
print("CSV gespeichert:", output_csv)

### Erstellen einer CSV mit allen Niederschlagsdaten von 2006 bis 2100

final_df = pd.concat(all_data, ignore_index=True)
final_df.to_csv(output_csv, index=False)
print("Erste Zeilen aus CSV:")
print(pd.read_csv(output_csv).head())


### Überblick über den neuen Datensatz

print(df.head())

print(df["temp_day"].describe())
