==> 2_code_datenaufbereitung/10_RCP_merge_discharge_clean.ipynb <==
## CSV-Dateien der RCPs zusammenfügen

import pandas as pd

# Einlesen und Spalten umbennen
rcp26_tas = pd.read_csv("https://raw.githubusercontent.com/emorl/projektstudie-murgtal/main/1_daten/rcp26pr/murg_pr_rcp26_mean.csv").rename(columns={"precip_mm_day": "pr_day_rcp26"})
rcp85_tas = pd.read_csv("https://raw.githubusercontent.com/emorl/projektstudie-murgtal/main/1_daten/rcp26temp/murg_tas_rcp26_mean.csv").rename(columns={"temp_day": "tas_day_rcp26"})
rcp26_pr   = pd.read_csv("https://raw.githubusercontent.com/emorl/projektstudie-murgtal/main/1_daten/rcp85pr/murg_pr_rcp85_mean.csv").rename(columns={"precip_mm_day": "pr_day_rcp85"})
rcp85_pr   = pd.read_csv("https://raw.githubusercontent.com/emorl/projektstudie-murgtal/main/1_daten/rcp85pr/murg_pr_rcp85_mean.csv").rename(columns={"temp_day": "tas_day_rcp85"})

# date als Datetime-Objekt
for df in [rcp26_tas, rcp85_tas, rcp26_pr, rcp85_pr]:
    df["date"] = pd.to_datetime(df["date"])

# Zusammenführen über date-Spalte
merged_df = rcp26_tas.merge(rcp85_tas, on="date") \
                      .merge(rcp26_pr, on="date") \
                      .merge(rcp85_pr, on="date")

#lokal speichern
output_csv = "hier-lokalen-Pfad-definieren/1_daten/tas_pr_merge.csv" 
merged_df.to_csv(output_csv, index=False)

print(merged_df.head())


# Abflussdaten vorbereiten und neu speichern

import pandas as pd

# CSV ab Zeile 7 einlesen mit korrekter Kodierung, da erste 6 Zeilen Projektbeschreibung sind
df = pd.read_csv(
    "https://raw.githubusercontent.com/emorl/projektstudie-murgtal/main/1_daten/discharge/Hydrologische-Landespegel-2025-12-22.csv",
    encoding="ISO-8859-1",  
    sep=",",                
    skiprows=6,             
    decimal=",",            
    quotechar='"'          
)

# Anfang anzeigen lassen
print(df.head())

# Datum umwandeln
df["Datum / Uhrzeit"] = pd.to_datetime(df["Datum / Uhrzeit"], errors="coerce")

# Nur Datum und Abflusswerte behalten
df = df[["Datum / Uhrzeit", "Wert"]].rename(columns={
    "Datum / Uhrzeit": "date",
    "Wert": "flow_m3_s"
})

print(df.head())


#Anzeigen lassen
df.describe()

# Neue Datei speichern
out_csv = "hier-lokalen-Pfad-definieren/1_daten/discharge/discharge_clean.csv"
df.to_csv(out_csv, index=False, encoding="utf-8")
