==> 2_code_datenaufbereitung/12_Neuformatierung_nach_Biaskorrektur.ipynb <==
# ReKlies Layout anpassen 

Für die prediction benötigt das Modell Datensätze, die genau gleich aufgebaut sind, wie die, die zum Training verwendet wurden. <br>

# libraries importieren
import pandas as pd
import os

data = pd.read_csv("https://raw.githubusercontent.com/emorl/projektstudie-murgtal/main/1_daten/future_bias_corrected_merge.csv")

# Datensatz RCP 2.6 

df_rcp26 = data.copy()

# Datum
df_rcp26['date'] = (
    pd.to_datetime(df_rcp26['date'], format='%Y-%m-%d %H:%M:%S')
      .dt.strftime('%d/%m/%Y')
)

# Tabellenköpfe umbenennen
df_rcp26 = df_rcp26.rename(columns={
    'tas_day_rcp26': 't2m_mean',
    'pr_day_rcp26': 'tp_sum'
})

# RCP 8.5 Spalten entfernen
df_rcp26 = df_rcp26.drop(columns=['tas_day_rcp85', 'pr_day_rcp85'], errors='ignore')

# neu anordnen
desired_order = ['date', 't2m_mean', 'tp_sum']
other_cols = [c for c in df_rcp26.columns if c not in desired_order]
df_rcp26 = df_rcp26[desired_order + other_cols]



# Datensatz RCP 8.5

df_rcp85 = data.copy()

# Datum 
df_rcp85['date'] = (
    pd.to_datetime(df_rcp85['date'], format='%Y-%m-%d %H:%M:%S')
      .dt.strftime('%d/%m/%Y')
)

# RCP 2.6 Spalten entfernen
df_rcp85 = df_rcp85.drop(columns=['tas_day_rcp26', 'pr_day_rcp26'], errors='ignore')

# Tabellenkopf umbenennen
df_rcp85 = df_rcp85.rename(columns={
    'tas_day_rcp85': 't2m_mean',
    'pr_day_rcp85': 'tp_sum'
})

# neu anordnen
desired_order = ['date', 't2m_mean', 'tp_sum']
other_cols = [c for c in df_rcp85.columns if c not in desired_order]
df_rcp85 = df_rcp85[desired_order + other_cols]


# beide Datensätze lokal speichern
df_rcp26.to_csv("C:/Users/henry/Documents/A git/projektstudie-murgtal/1_daten/future_bias_corrected_rcp26.csv", index=False)
df_rcp85.to_csv("C:/Users/henry/Documents/A git/projektstudie-murgtal/1_daten/future_bias_corrected_rcp85.csv", index=False)
